% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalSmoothFunctions.R
\name{CalculateLOOCV}
\alias{CalculateLOOCV}
\title{Calculate Leave-One-Out Cross Validation}
\usage{
CalculateLOOCV(x, y, lambda)
}
\arguments{
\item{x}{Vector of data points.}

\item{y}{Vector of observations}

\item{lambda}{Smoothing parameter.}
}
\value{
Single numeric value representing the mean sqaure CV error.
}
\description{
Returns the value of the leave-one-out cross validation score for a given
  data set and smoothing parameter when a P-Spline fit is used.
}
