% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffMatrix.R
\name{CPerm}
\alias{CPerm}
\title{Cyclic Permuatation}
\usage{
CPerm(x, i = 1)
}
\arguments{
\item{x}{Vector to be cycled through}

\item{i}{Number of cyclic permutations}
}
\value{
Vector with all elements moved \code{i} spaces to the left with
  end elements wrapping around.
}
\description{
Get cyclic permutation of vector. Can can the function multiple times
  through the use of the \code{i} argument.
}
