% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EgData129}
\alias{EgData129}
\title{Random variation around a Sine curve (n = 129)}
\format{An object of class \code{list} of length 2.}
\usage{
EgData129
}
\description{
The \code{x} values are equally spaced points at
  0, \eqn{pi / (n - 1)}, ..., \eqn{2 pi}. whereas the \code{y} values were
  generated by adding draws from a random normal distribution
  (\eqn{s.d. = 0.1}) around a Sine curve.
}
\keyword{datasets}
