% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSplineFunctions.R
\name{FitPSpline}
\alias{FitPSpline}
\title{Fit Natural cubic spline with data points as knots (P-Spline)}
\usage{
FitPSpline(x, y, lambda)
}
\arguments{
\item{x}{Vector of data points.}

\item{y}{Vector of observations}

\item{lambda}{Smoothing parameter.}
}
\value{
Named list with the following components:

\code{AlphaEst} is a vector of estimated coefficients of the
  regression model.

\code{YEst} is a vector of fitted values.

\code{VarYEst} is the dispersion matrix of the fitted value estimates.
}
\description{
This will utilise the B-Spline functions in this package to fit a
  P-Spline. That is, a cubic B-spline with interior and exterior knots
  defined by each data point.
}
