% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplineFunctions.R
\name{GetBasis}
\alias{GetBasis}
\title{Get Basis functions}
\usage{
GetBasis(x, deg, KKnots, i)
}
\arguments{
\item{x}{Range of values to define the function over.}

\item{deg}{Degree of the desired B-Spline.}

\item{KKnots}{Augmented knot set.}

\item{i}{Current level of basis function.}
}
\value{
The \eqn{i}-th basis function \eqn{\phi} of level \code{deg}
  from B-Spline generative procedure
}
\description{
Retrieves the \eqn{i}-th basis function mapping for some given range
  \code{x}, degree and augmented knot set by means of recursion.
}
