% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplineFunctions.R
\name{GetBSpline}
\alias{GetBSpline}
\title{Get B-Spline Matrix}
\usage{
GetBSpline(x, deg = 3, IntKnots, ExtKnots)
}
\arguments{
\item{x}{Range of values to define the function over.}

\item{deg}{Degree of the desired B-Spline.}

\item{IntKnots}{Interior knots that partially define the B-Spline.}

\item{ExtKnots}{Exterior knots, often \code{ExtKnots = c(min(x),max(x))}.}
}
\value{
Matrix where \eqn{i,j}-th entry corresponds to \eqn{j}-th basis
function evaluated at \eqn{i}-th data point.
}
\description{
Generates B-Spline functions over some parameters and places such functions
into columns of a \eqn{n} by \eqn{(m + deg + 1)} matrix
}
