% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffMatrix.R
\name{GetDiffMatrix}
\alias{GetDiffMatrix}
\title{Get Difference Matrix}
\usage{
GetDiffMatrix(nr, ord = 2)
}
\arguments{
\item{nr}{Number of rows the difference matrix will have.}

\item{ord}{Order of the difference matrix.}
}
\value{
A difference matrix of order \code{ord} with dimensions \code{nr}
  by \eqn{\code{nr} + \code{ord}}
}
\description{
Calculates the difference matrix of order 2 (ISSUE -- Only gets difference
  matrix of order 2 for now).
}
