% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalSmoothFunctions.R
\name{MinimiseLOOCV}
\alias{MinimiseLOOCV}
\title{Minimised Leave-One-Out Cross Validation}
\usage{
MinimiseLOOCV(x, y, Lambdas, toPlot = T, ...)
}
\arguments{
\item{x}{Vector of data points.}

\item{y}{Vector of observations}

\item{Lambdas}{Vector of smoothing paramaeters to test.}

\item{toPlot}{Logical. Should the plot of CV error against
\code{Lambdas} appear?}

\item{...}{Optional graphical parameters.}
}
\description{
Finds the value from a given range that minimised the mean sqaure cross
  validation error.
}
