% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSplineFunctions.R
\name{PlotBSpline}
\alias{PlotBSpline}
\title{Plot B-Spline functions}
\usage{
PlotBSpline(x, B, lty = 1:ncol(B), col = 1:ncol(B))
}
\arguments{
\item{x}{Same \code{x} used in \code{GetBSpline(x, ...)}.}

\item{B}{Output from \code{GetBSpline()} call.}

\item{lty}{Line type. Can be single value or range of values.}

\item{col}{Line colour. Can be single value or range of values.}
}
\description{
Takes some B-Spline basis matrix from \code{GetBSpline()} and plots the
output. *Note* \code{x} must match the \code{x} used in
\code{GetBSpline(x, ...)}.
}
