% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalMAP.R
\name{PlotPostDensity}
\alias{PlotPostDensity}
\title{Plot Posterior Density for Smoothing Parameter in PSpline}
\usage{
PlotPostDensity(x, y, Lambdas, PriorDens, ...)
}
\arguments{
\item{y}{Vector of observations.}

\item{Lambdas}{Range of smoothing parameters to plot / scan.}

\item{PriorDens}{Prior density of penalisation parameter.}
}
\value{
Value of smoothing parameter that reuturned minimal density.
}
\description{
Plots the posterior density (to be used with MAP) when a natural cubic
  spline and second order differencing matrix approximation is used.
}
