% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSplineFunctions.R
\name{PlotPSplineFit}
\alias{PlotPSplineFit}
\title{Plot P-Spline Fit to Data}
\usage{
PlotPSplineFit(x, y, lambda, CI = 0.95, ...)
}
\arguments{
\item{x}{Vector of data points.}

\item{y}{Vector of observations}

\item{lambda}{Smoothing parameter.}

\item{CI}{Given confidence level.}

\item{...}{Optional graphical parameters.}
}
\description{
Plot given data with curve showing fitted values as well as overlaying
  confidence intervals for given probability.
}
