% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalMAP.R
\name{postDensity}
\alias{postDensity}
\title{Posterior Density of Lambda}
\usage{
postDensity(lambda, B, D, a, b, y, PriorDens = 1)
}
\arguments{
\item{lambda}{Parameter of focus.}

\item{B}{Design matrix obtained via natural cubic spline.}

\item{D}{Difference matrix used in penalty term approximation.}

\item{a, b}{Prior information.}

\item{y}{Vector of observations.}

\item{PriorDens}{Prior density of penalisation parameter.}
}
\description{
Uses the equation given in "Efficient and automatic methods for flexible
  regression on spatiotemporal data, with applications to groundwater
  monitoring" by Evers et al. (2015).
}
