% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSplineFunctions.R
\name{PredictPSpline}
\alias{PredictPSpline}
\title{Prediction via P-Spline Model}
\usage{
PredictPSpline(x, y, lambda, XNew)
}
\arguments{
\item{x}{Vector of data points.}

\item{y}{Vector of observations}

\item{lambda}{Smoothing parameter.}

\item{XNew}{A single data point, for which to predict a \code{y} value.
(Need to update code to accept vectors of XNew).}
}
\value{
Single predicted \code{y} value. (See note above).
}
\description{
Predict a value for \eqn{y'}, given new value \eqn{x'} based on regression
model and parameter estimates found via P-Spline method.
}
