% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalMAP.R
\name{PlotPostDensity}
\alias{PlotPostDensity}
\title{Plot Posterior Density for Smoothing Parameter in PSpline}
\usage{
PlotPostDensity(x, y, a, b, Lambdas, PriorDens, add = F, ...)
}
\arguments{
\item{y}{Vector of observations.}

\item{a}{Prior information related to normal-inverse gamma distribution.}

\item{b}{Prior information related to normal-inverse gamma distribution.}

\item{Lambdas}{Range of smoothing parameters to plot / scan.}

\item{PriorDens}{Prior density of penalisation parameter.}

\item{add}{Logical. If true, plot is added to currently open graphical
space.}
}
\value{
Value of smoothing parameter that reuturned minimal density.
}
\description{
Plots the posterior density (to be used with MAP) when a natural cubic
  spline and second order differencing matrix approximation is used.
}
